+++ src/cspace/cap/asid_control.rs
@@ -11,3 +11,8 @@ impl cap_t {
         cap
     }
 }
+
+#[inline]
+pub fn cap_asid_control_cap_new() -> cap_t {
+    cap_t::new_asid_control_cap()
+}
