+++ src/cspace/cap/frame.rs
@@ -73,3 +73,53 @@ impl cap_t {
         self.words[0] |= (v64 >> 0) & 0x7fffffffffusize;
     }
 }
+
+
+#[inline]
+pub fn cap_frame_cap_new(capFMappedASID: usize, capFBasePtr: usize, capFSize: usize, capFVMRights: usize, capFIsDevice: usize, capFMappedAddress: usize) -> cap_t {
+    cap_t::new_frame_cap(capFMappedASID, capFBasePtr, capFSize, capFVMRights, capFIsDevice, capFMappedAddress)
+}
+
+#[inline]
+pub fn cap_frame_cap_get_capFMappedASID(cap: &cap_t) -> usize {
+    cap.get_frame_mapped_asid()
+}
+
+#[inline]
+pub fn cap_frame_cap_set_capFMappedASID(cap: &mut cap_t, v64: usize) {
+    cap.set_frame_mapped_asid(v64)
+}
+
+#[inline]
+pub fn cap_frame_cap_get_capFBasePtr(cap: &cap_t) -> usize {
+    cap.get_frame_base_ptr()
+}
+
+#[inline]
+pub fn cap_frame_cap_get_capFSize(cap: &cap_t) -> usize {
+    cap.get_frame_size()
+}
+
+#[inline]
+pub fn cap_frame_cap_get_capFVMRights(cap: &cap_t) -> usize {
+    cap.get_frame_vm_rights()
+}
+
+#[inline]
+pub fn cap_frame_cap_set_capFVMRights(cap: &mut cap_t, v64: usize) {
+    cap.set_frame_vm_rights(v64)
+}
+
+#[inline]
+pub fn cap_frame_cap_get_capFIsDevice(cap: &cap_t) -> usize {
+    cap.get_frame_is_device()
+}
+
+#[inline]
+pub fn cap_frame_cap_get_capFMappedAddress(cap: &cap_t) -> usize {
+    cap.get_frame_mapped_address()
+}
+#[inline]
+pub fn cap_frame_cap_set_capFMappedAddress(cap: &mut cap_t, v64: usize) {
+    cap.set_frame_mapped_address(v64)
+}
