+++ src/cspace/cap/irq_control.rs
@@ -12,3 +12,8 @@ impl cap_t {
         cap
     }
 }
+
+#[inline]
+pub fn cap_irq_control_cap_new() -> cap_t {
+    cap_t::new_irq_control_cap()
+}
