+++ src/cspace/cap/null.rs
@@ -10,3 +10,8 @@ impl cap_t {
         cap
     }
 }
+
+#[inline]
+pub fn cap_null_cap_new() -> cap_t {
+    cap_t::new_null_cap()
+}
