+++ src/kernel/syscall.rs
@@ -12,9 +12,6 @@ use crate::{
         notification::receiveSignal,
         objecttype::decodeInvocation,
         structure_gen::{
-            cap_endpoint_cap_get_capCanReceive, cap_notification_cap_get_capNtfnCanReceive,
-            cap_notification_cap_get_capNtfnPtr, cap_reply_cap_get_capReplyCanGrant,
-            cap_reply_cap_get_capReplyMaster, cap_reply_cap_get_capTCBPtr,
             lookup_fault_missing_capability_new, notification_ptr_get_ntfnBoundTCB,
             seL4_Fault_CapFault_new, seL4_Fault_UserException_new, thread_state_get_tsType,
         },
