+++ src/kernel/vspace.rs
@@ -21,16 +21,6 @@ use crate::{
         cap::{cteInsert, ensureEmptySlot, ensureNoChildren, isFinalCapability},
         objecttype::cap_get_capPtr,
         structure_gen::{
-            cap_asid_pool_cap_get_capASIDBase, cap_asid_pool_cap_get_capASIDPool,
-            cap_frame_cap_get_capFBasePtr, cap_frame_cap_get_capFIsDevice,
-            cap_frame_cap_get_capFMappedASID, cap_frame_cap_get_capFMappedAddress,
-            cap_frame_cap_get_capFSize, cap_frame_cap_get_capFVMRights,
-            cap_frame_cap_set_capFMappedASID, cap_frame_cap_set_capFMappedAddress,
-            cap_page_table_cap_get_capPTBasePtr,
-            cap_page_table_cap_get_capPTIsMapped, cap_page_table_cap_get_capPTMappedASID,
-            cap_page_table_cap_get_capPTMappedAddress,
-            cap_page_table_cap_ptr_set_capPTIsMapped, cap_page_table_cap_set_capPTIsMapped,
-            cap_page_table_cap_set_capPTMappedASID, cap_page_table_cap_set_capPTMappedAddress,
             lookup_fault_invalid_root_new, lookup_fault_missing_capability_new,
             pte_ptr_get_execute, pte_ptr_get_ppn, pte_ptr_get_read, pte_ptr_get_valid,
             pte_ptr_get_write, seL4_Fault_VMFault_new,
