+++ src/object/objecttype.rs
@@ -44,28 +44,6 @@ use super::{
         unbindNotification,
     },
     structure_gen::{
-        cap_asid_pool_cap_get_capASIDBase, cap_asid_pool_cap_get_capASIDPool,
-        cap_cnode_cap_get_capCNodePtr, cap_cnode_cap_get_capCNodeRadix,
-        cap_cnode_cap_set_capCNodeGuard, cap_cnode_cap_set_capCNodeGuardSize,
-        cap_endpoint_cap_get_capCanGrant, cap_endpoint_cap_get_capCanGrantReply,
-        cap_endpoint_cap_get_capCanReceive, cap_endpoint_cap_get_capCanSend,
-        cap_endpoint_cap_get_capEPBadge, cap_endpoint_cap_get_capEPPtr,
-        cap_endpoint_cap_set_capCanGrant, cap_endpoint_cap_set_capCanGrantReply,
-        cap_endpoint_cap_set_capCanReceive, cap_endpoint_cap_set_capCanSend,
-        cap_endpoint_cap_set_capEPBadge, cap_frame_cap_get_capFBasePtr,
-        cap_frame_cap_get_capFIsDevice, cap_frame_cap_get_capFMappedASID,
-        cap_frame_cap_get_capFMappedAddress, cap_frame_cap_get_capFSize,
-        cap_frame_cap_get_capFVMRights, cap_frame_cap_set_capFMappedASID,
-        cap_frame_cap_set_capFMappedAddress, cap_frame_cap_set_capFVMRights,
-        cap_irq_handler_cap_get_capIRQ, cap_notification_cap_get_capNtfnBadge,
-        cap_notification_cap_get_capNtfnCanReceive, cap_notification_cap_get_capNtfnCanSend,
-        cap_notification_cap_get_capNtfnPtr,
-        cap_notification_cap_set_capNtfnBadge, cap_notification_cap_set_capNtfnCanReceive,
-        cap_notification_cap_set_capNtfnCanSend,
-        cap_page_table_cap_get_capPTBasePtr, cap_page_table_cap_get_capPTIsMapped,
-        cap_page_table_cap_get_capPTMappedASID, cap_page_table_cap_get_capPTMappedAddress, cap_reply_cap_get_capReplyCanGrant,
-        cap_reply_cap_get_capReplyMaster, cap_reply_cap_get_capTCBPtr,
-        cap_reply_cap_set_capReplyCanGrant, cap_thread_cap_get_capTCBPtr,
         cap_zombie_cap_get_capZombiePtr, ZombieType_ZombieTCB, Zombie_new,
     },
     tcb::decodeTCBInvocation,
