+++ src/object/structure_gen.rs
@@ -37,294 +37,6 @@ pub fn cap_get_max_free_index(cap: &cap_t) -> usize {
 
 
 #[inline]
-pub fn cap_endpoint_cap_get_capEPBadge(cap: &cap_t) -> usize {
-    let mut ret: usize;
-    assert!(((cap.words[0] >> 59) & 0x1f) == cap_tag_t::cap_endpoint_cap as usize);
-
-    ret = (cap.words[1] & 0xffffffffffffffffusize) >> 0;
-    /* Possibly sign extend */
-    if unlikely(!!(false && (ret & (1usize << (38))) != 0)) {
-        ret |= 0x0;
-    }
-    return ret;
-}
-
-#[inline]
-pub fn cap_endpoint_cap_set_capEPBadge(cap: &mut cap_t, v64: usize) {
-    assert!(((cap.words[0] >> 59) & 0x1f) == cap_tag_t::cap_endpoint_cap as usize);
-    /* fail if user has passed bits that we will override */
-    assert!(
-        (((!0xffffffffffffffffusize >> 0) | 0x0) & v64)
-            == (if false && (v64 & (1usize << (38))) != 0 {
-                0x0
-            } else {
-                0
-            })
-    );
-
-    cap.words[1] &= !0xffffffffffffffffusize;
-    cap.words[1] |= (v64 << 0) & 0xffffffffffffffffusize;
-}
-
-#[inline]
-pub fn cap_endpoint_cap_get_capCanGrantReply(cap: & cap_t) -> usize {
-        let mut ret: usize;
-        assert!(((cap.words[0] >> 59) & 0x1f) == cap_tag_t::cap_endpoint_cap as usize);
-
-        ret = (cap.words[0] & 0x400000000000000usize) >> 58;
-        /* Possibly sign extend */
-        if unlikely(!!(false && (ret & (1usize << (38))) != 0)) {
-            ret |= 0x0;
-        }
-        return ret;
-}
-
-#[inline]
-pub fn cap_endpoint_cap_set_capCanGrantReply(cap: &mut cap_t, v64: usize) {
-    assert!(((cap.words[0] >> 59) & 0x1f) == cap_tag_t::cap_endpoint_cap as usize);
-    /* fail if user has passed bits that we will override */
-    assert!(
-        (((!0x400000000000000usize >> 58) | 0x0) & v64)
-            == (if false && (v64 & (1usize << (38))) != 0 {
-                0x0
-            } else {
-                0
-            })
-    );
-    cap.words[0] &= !0x400000000000000usize;
-    cap.words[0] |= (v64 << 58) & 0x400000000000000usize;
-}
-
-#[inline]
-pub fn cap_endpoint_cap_get_capCanGrant(cap: &cap_t) -> usize {
-    let mut ret: usize;
-    assert!(((cap.words[0] >> 59) & 0x1f) == cap_tag_t::cap_endpoint_cap as usize);
-
-    ret = (cap.words[0] & 0x200000000000000usize) >> 57;
-    /* Possibly sign extend */
-    if unlikely(!!(false && (ret & (1usize << (38))) != 0)) {
-        ret |= 0x0;
-    }
-    return ret;
-}
-#[inline]
-pub fn cap_endpoint_cap_set_capCanGrant(cap: &mut cap_t, v64: usize) {
-    assert!(((cap.words[0] >> 59) & 0x1f) == cap_tag_t::cap_endpoint_cap as usize);
-    /* fail if user has passed bits that we will override */
-    assert!(
-        (((!0x200000000000000usize >> 57) | 0x0) & v64)
-            == (if false && (v64 & (1usize << (38))) != 0 {
-                0x0
-            } else {
-                0
-            })
-    );
-
-    cap.words[0] &= !0x200000000000000usize;
-    cap.words[0] |= (v64 << 57) & 0x200000000000000usize;
-}
-
-#[inline]
-pub fn cap_endpoint_cap_get_capCanReceive(cap: &cap_t) -> usize {
-    let mut ret: usize;
-    assert!(((cap.words[0] >> 59) & 0x1f) == cap_tag_t::cap_endpoint_cap as usize);
-
-    ret = (cap.words[0] & 0x100000000000000usize) >> 56;
-    /* Possibly sign extend */
-    if unlikely(!!(false && (ret & (1usize << (38))) != 0)) {
-        ret |= 0x0;
-    }
-    return ret;
-}
-
-#[inline]
-pub fn cap_endpoint_cap_set_capCanReceive(cap: &mut cap_t, v64: usize) {
-    assert!(((cap.words[0] >> 59) & 0x1f) == cap_tag_t::cap_endpoint_cap as usize);
-    /* fail if user has passed bits that we will override */
-    assert!(
-        (((!0x100000000000000usize >> 56) | 0x0) & v64)
-            == (if false && (v64 & (1usize << (38))) != 0 {
-                0x0
-            } else {
-                0
-            })
-    );
-
-    cap.words[0] &= !0x100000000000000usize;
-    cap.words[0] |= (v64 << 56) & 0x100000000000000usize;
-}
-
-#[inline]
-pub fn cap_endpoint_cap_get_capCanSend(cap: &cap_t) -> usize {
-    let mut ret: usize;
-    assert!(((cap.words[0] >> 59) & 0x1f) == cap_tag_t::cap_endpoint_cap as usize);
-
-    ret = (cap.words[0] & 0x80000000000000usize) >> 55;
-    /* Possibly sign extend */
-    if unlikely(!!(false && (ret & (1usize << (38))) != 0)) {
-        ret |= 0x0;
-    }
-    return ret;
-}
-
-#[inline]
-pub fn cap_endpoint_cap_set_capCanSend(cap: &mut cap_t, v64: usize) {
-    assert!(((cap.words[0] >> 59) & 0x1f) == cap_tag_t::cap_endpoint_cap as usize);
-    /* fail if user has passed bits that we will override */
-    assert!(
-        (((!0x80000000000000usize >> 55) | 0x0) & v64)
-            == (if false && (v64 & (1usize << (38))) != 0 {
-                0x0
-            } else {
-                0
-            })
-    );
-
-    cap.words[0] &= !0x80000000000000usize;
-    cap.words[0] |= (v64 << 55) & 0x80000000000000usize;
-}
-
-#[inline]
-pub fn cap_endpoint_cap_get_capEPPtr(cap: &cap_t) -> usize {
-    let mut ret: usize;
-    assert!(((cap.words[0] >> 59) & 0x1f) == cap_tag_t::cap_endpoint_cap as usize);
-
-    ret = (cap.words[0] & 0x7fffffffffusize) << 0;
-    /* Possibly sign extend */
-    if likely(!!(true && (ret & (1usize << (38))) != 0)) {
-        ret |= 0xffffff8000000000;
-    }
-    ret
-}
-
-#[inline]
-pub fn cap_cnode_cap_get_capCNodeGuard(cap: &cap_t) -> usize {
-    let mut ret: usize;
-    assert!(((cap.words[0] >> 59) & 0x1f) == cap_tag_t::cap_cnode_cap as usize);
-
-    ret = (cap.words[1] & 0xffffffffffffffffusize) >> 0;
-    /* Possibly sign extend */
-    if unlikely(!!(false && (ret & (1usize << (38))) != 0)) {
-        ret |= 0x0;
-    }
-    return ret;
-}
-
-#[inline]
-pub fn cap_cnode_cap_set_capCNodeGuard(cap: &mut cap_t, v64: usize) {
-    assert!(((cap.words[0] >> 59) & 0x1f) == cap_tag_t::cap_cnode_cap as usize);
-    /* fail if user has passed bits that we will override */
-    assert!(
-        (((!0xffffffffffffffffusize >> 0) | 0x0) & v64)
-            == (if false && (v64 & (1usize << (38))) != 0 {
-                0x0
-            } else {
-                0
-            })
-    );
-
-    cap.words[1] &= !0xffffffffffffffffusize;
-    cap.words[1] |= (v64 << 0) & 0xffffffffffffffffusize;
-}
-
-#[inline]
-pub fn cap_cnode_cap_get_capCNodeGuardSize(cap: &cap_t) -> usize {
-    let mut ret: usize;
-    assert!(((cap.words[0] >> 59) & 0x1f) == cap_tag_t::cap_cnode_cap as usize);
-
-    ret = (cap.words[0] & 0x7e0000000000000usize) >> 53;
-    /* Possibly sign extend */
-    if unlikely(!!(false && (ret & (1usize << (38))) != 0)) {
-        ret |= 0x0;
-    }
-    return ret;
-}
-
-#[inline]
-pub fn cap_cnode_cap_set_capCNodeGuardSize(cap: &mut cap_t, v64: usize) {
-    assert!(((cap.words[0] >> 59) & 0x1f) == cap_tag_t::cap_cnode_cap as usize);
-    /* fail if user has passed bits that we will override */
-    assert!(
-        (((!0x7e0000000000000usize >> 53) | 0x0) & v64)
-            == (if false && (v64 & (1usize << (38))) != 0 {
-                0x0
-            } else {
-                0
-            })
-    );
-
-    cap.words[0] &= !0x7e0000000000000usize;
-    cap.words[0] |= (v64 << 53) & 0x7e0000000000000usize;
-}
-
-#[inline]
-pub fn cap_cnode_cap_get_capCNodeRadix(cap: &cap_t) -> usize {
-    let mut ret;
-    assert!(((cap.words[0] >> 59) & 0x1f) == cap_tag_t::cap_cnode_cap as usize);
-
-    ret = (cap.words[0] & 0x1f800000000000usize) >> 47;
-    /* Possibly sign extend */
-    if unlikely(!!(false && (ret & (1usize << (38))) != 0)) {
-        ret |= 0x0;
-    }
-    return ret;
-}
-
-#[inline]
-pub fn cap_cnode_cap_get_capCNodePtr(cap: &cap_t) -> usize {
-    let mut ret: usize;
-    assert!(((cap.words[0] >> 59) & 0x1f) == cap_tag_t::cap_cnode_cap as usize);
-
-    ret = (cap.words[0] & 0x3fffffffffusize) << 1;
-    /* Possibly sign extend */
-    if (ret & (1 << 38)) != 0 {
-        ret |= 0xffffff8000000000;
-    }
-    ret
-}
-
-#[inline]
-pub fn isArchCap(cap: &cap_t) -> bool {
-    cap_get_capType(cap) % 2 != 0
-}
-
-
-#[inline]
-pub fn cap_zombie_cap_get_capZombieID(cap: &cap_t) -> usize {
-    let  ret;
-    assert!(((cap.words[0] >> 59) & 0x1f) == cap_tag_t::cap_zombie_cap as usize);
-
-    ret = (cap.words[1] & 0xffffffffffffffffusize) >> 0;
-    ret
-}
-
-#[inline]
-pub fn cap_zombie_cap_set_capZombieID(cap: &mut cap_t, v64: usize) {
-    assert!(((cap.words[0] >> 59) & 0x1f) == cap_tag_t::cap_zombie_cap as usize);
-    /* fail if user has passed bits that we will override */
-    assert!(
-        (((!0xffffffffffffffffusize >> 0) | 0x0) & v64)
-            == (if false && (v64 & (1usize << (38))) != 0 {
-                0x0
-            } else {
-                0
-            })
-    );
-
-    cap.words[1] &= !0xffffffffffffffffusize;
-    cap.words[1] |= (v64 << 0) & 0xffffffffffffffffusize;
-}
-
-#[inline]
-pub fn cap_zombie_cap_get_capZombieType(cap: &cap_t) -> usize {
-    let ret;
-    assert!(((cap.words[0] >> 59) & 0x1f) == cap_tag_t::cap_zombie_cap as usize);
-
-    ret = (cap.words[0] & 0x7fusize) >> 0;
-    return ret;
-}
-
-#[inline]
 pub fn Zombie_new(number: usize, _type: usize, ptr: usize) -> cap_t {
     let mask: usize;
     if _type == ZombieType_ZombieTCB {
@@ -363,118 +75,6 @@ pub fn cap_zombie_cap_set_capZombieNumber(cap: &mut cap_t, n: usize) {
 
 
 #[inline]
-pub fn cap_page_table_cap_get_capPTMappedASID(cap: &cap_t) -> usize {
-    let ret = (cap.words[1] & 0xffff000000000000usize) >> 48;
-    ret
-}
-
-#[inline]
-pub fn cap_page_table_cap_set_capPTMappedASID(cap: &mut cap_t, v64: usize) {
-    cap.words[1] &= !0xffff000000000000usize;
-    cap.words[1] |= (v64 << 48) & 0xffff000000000000usize;
-}
-
-#[inline]
-pub fn cap_page_table_cap_get_capPTBasePtr(cap: &cap_t) -> usize {
-    let mut ret = (cap.words[1] & 0xfffffffffe00usize) >> 9;
-    if (ret & (1usize << (38))) != 0 {
-        ret |= 0xffffff8000000000;
-    }
-    ret
-}
-
-#[inline]
-pub fn cap_page_table_cap_get_capPTIsMapped(cap: &cap_t) -> usize {
-    let ret = (cap.words[0] & 0x8000000000usize) >> 39;
-    ret
-}
-
-#[inline]
-pub fn cap_page_table_cap_set_capPTIsMapped(cap: &mut cap_t, v64: usize) {
-    cap.words[0] &= !0x8000000000usize;
-    cap.words[0] |= (v64 << 39) & 0x8000000000usize;
-}
-
-#[inline]
-pub fn cap_page_table_cap_ptr_set_capPTIsMapped(cap: &mut cap_t, v64: usize) {
-    cap.words[0] &= !0x8000000000usize;
-    cap.words[0] |= (v64 << 39) & 0x8000000000usize;
-}
-#[inline]
-pub fn cap_page_table_cap_get_capPTMappedAddress(cap: &cap_t) -> usize {
-    let mut ret = (cap.words[0] & 0x7fffffffffusize) << 0;
-    if (ret & (1usize << (38))) != 0 {
-        ret |= 0xffffff8000000000;
-    }
-    ret
-}
-
-#[inline]
-pub fn cap_page_table_cap_set_capPTMappedAddress(cap: &mut cap_t, v64: usize) {
-    cap.words[0] &= !0x7fffffffffusize;
-    cap.words[0] |= (v64 >> 0) & 0x7fffffffffusize;
-}
-
-#[inline]
-pub fn cap_frame_cap_get_capFMappedASID(cap: &cap_t) -> usize {
-    let ret = (cap.words[1] & 0xffff000000000000usize) >> 48;
-    ret
-}
-
-#[inline]
-pub fn cap_frame_cap_set_capFMappedASID(cap: &mut cap_t, v64: usize) {
-    cap.words[1] &= !0xffff000000000000usize;
-    cap.words[1] |= (v64 << 48) & 0xffff000000000000usize;
-}
-
-#[inline]
-pub fn cap_frame_cap_get_capFBasePtr(cap: &cap_t) -> usize {
-    let mut ret = (cap.words[1] & 0xfffffffffe00usize) >> 9;
-    if (ret & (1usize << (38))) != 0 {
-        ret |= 0xffffff8000000000;
-    }
-    ret
-}
-
-#[inline]
-pub fn cap_frame_cap_get_capFSize(cap: &cap_t) -> usize {
-    let ret = (cap.words[0] & 0x600000000000000usize) >> 57;
-    ret
-}
-
-#[inline]
-pub fn cap_frame_cap_get_capFVMRights(cap: &cap_t) -> usize {
-    let ret = (cap.words[0] & 0x180000000000000usize) >> 55;
-    ret
-}
-
-#[inline]
-pub fn cap_frame_cap_set_capFVMRights(cap: &mut cap_t, v64: usize) {
-    cap.words[0] &= !0x180000000000000usize;
-    cap.words[0] |= (v64 << 55) & 0x180000000000000usize;
-}
-
-#[inline]
-pub fn cap_frame_cap_get_capFIsDevice(cap: &cap_t) -> usize {
-    let ret = (cap.words[0] & 0x40000000000000usize) >> 54;
-    ret
-}
-
-#[inline]
-pub fn cap_frame_cap_get_capFMappedAddress(cap: &cap_t) -> usize {
-    let mut ret = (cap.words[0] & 0x7fffffffffusize) << 0;
-    if (ret & (1usize << 38)) != 0 {
-        ret |= 0xffffff8000000000;
-    }
-    ret
-}
-#[inline]
-pub fn cap_frame_cap_set_capFMappedAddress(cap: &mut cap_t, v64: usize) {
-    cap.words[0] &= !0x7fffffffffusize;
-    cap.words[0] |= (v64 >> 0) & 0x7fffffffffusize;
-}
-
-#[inline]
 pub fn pte_ptr_get_ppn(pte_ptr: *const pte_t) -> usize {
     unsafe {
         let ret = ((*pte_ptr).words[0] & 0x3f_ffff_ffff_fc00usize) >> 10;
@@ -511,21 +111,6 @@ pub fn pte_ptr_get_valid(pte_ptr: *const pte_t) -> usize {
         let ret = ((*pte_ptr).words[0] & 0x1usize) >> 0;
         ret
     }
-}
-
-#[inline]
-pub fn cap_asid_pool_cap_get_capASIDBase(cap: &cap_t) -> usize {
-    let ret = ((*cap).words[0] & 0x7fff80000000000usize) >> 43;
-    ret
-}
-
-#[inline]
-pub fn cap_asid_pool_cap_get_capASIDPool(cap: &cap_t) -> usize {
-    let mut ret = (cap.words[0] & 0x1fffffffffusize) << 2;
-    if likely(!!(true && (ret & (1usize << (38))) != 0)) {
-        ret |= 0xffffff8000000000;
-    }
-    ret
 }
 
 #[inline]
@@ -681,40 +266,6 @@ pub fn endpoint_ptr_get_state(ptr: *const endpoint_t) -> usize {
 }
 
 #[inline]
-pub fn cap_reply_cap_get_capTCBPtr(cap: &cap_t) -> usize {
-    let ret = (cap).words[1] & 0xffffffffffffffffusize;
-    ret
-}
-
-#[inline]
-pub fn cap_reply_cap_get_capReplyCanGrant(cap: &cap_t) -> usize {
-    let ret = ((cap).words[0] & 0x2usize) >> 1;
-    ret
-}
-
-#[inline]
-pub fn cap_reply_cap_set_capReplyCanGrant(cap: &mut cap_t, v64: usize) {
-    (cap).words[0] &= !0x2usize;
-    (cap).words[0] |= (v64 << 1) & 0x2usize;
-}
-
-#[inline]
-pub fn cap_reply_cap_get_capReplyMaster(cap: &cap_t) -> usize {
-    let ret = ((cap).words[0] & 0x1usize) >> 0;
-    ret
-}
-
-#[inline]
-pub fn cap_thread_cap_get_capTCBPtr(cap: &cap_t) -> usize {
-    let mut ret: usize;
-    ret = ((cap).words[0] & 0x7fffffffffusize) << 0;
-    if (ret & (1usize << (38))) != 0 {
-        ret |= 0xffffff8000000000;
-    }
-    ret
-}
-
-#[inline]
 pub fn notification_ptr_get_ntfnBoundTCB(notification_ptr: *const notification_t) -> usize {
     let mut ret: usize;
     unsafe {
@@ -806,64 +357,6 @@ pub fn notification_ptr_set_state(ptr: *mut notification_t, v64: usize) {
         (*ptr).words[0] &= !0x3usize;
         (*ptr).words[0] |= (v64 >> 0) & 0x3usize;
     }
-}
-
-
-#[inline]
-pub fn cap_notification_cap_get_capNtfnBadge(cap: &cap_t) -> usize {
-    let ret = (cap).words[1] & 0xffffffffffffffffusize;
-    ret
-}
-
-#[inline]
-pub fn cap_notification_cap_set_capNtfnBadge(cap: &mut cap_t, v64: usize) {
-    (cap).words[1] &= !0xffffffffffffffffusize;
-    (cap).words[1] |= v64 & 0xffffffffffffffffusize;
-}
-
-#[inline]
-pub fn cap_notification_cap_get_capNtfnCanReceive(cap: &cap_t) -> usize {
-    let ret = ((cap).words[0] & 0x400000000000000usize) >> 58;
-    ret
-}
-
-#[inline]
-pub fn cap_notification_cap_set_capNtfnCanReceive(cap: &mut cap_t, v64: usize) {
-    (cap).words[0] &= !0x400000000000000usize;
-    (cap).words[0] |= (v64 << 58) & 0x400000000000000usize;
-}
-
-#[inline]
-pub fn cap_notification_cap_get_capNtfnCanSend(cap: &cap_t) -> usize {
-    let ret = ((cap).words[0] & 0x200000000000000usize) >> 57;
-    ret
-}
-
-#[inline]
-pub fn cap_notification_cap_set_capNtfnCanSend(cap: &mut cap_t, v64: usize) {
-    (cap).words[0] &= !0x200000000000000usize;
-    (cap).words[0] |= (v64 << 57) & 0x200000000000000usize;
-}
-
-#[inline]
-pub fn cap_notification_cap_get_capNtfnPtr(cap: &cap_t) -> usize {
-    let mut ret = (cap).words[0] & 0x7fffffffffusize;
-    if (ret & (1usize << (38))) != 0 {
-        ret |= 0xffffff8000000000;
-    }
-    ret
-}
-
-#[inline]
-pub fn cap_notification_cap_set_capNtfnPtr(cap: &mut cap_t, v64: usize) {
-    (cap).words[0] &= !0x7fffffffffusize;
-    (cap).words[0] |= v64 & 0x7fffffffffusize;
-}
-
-#[inline]
-pub fn cap_irq_handler_cap_get_capIRQ(cap: &cap_t) -> usize {
-    let ret = (cap.words[1] & 0xfffusize) >> 0;
-    ret
 }
 
 #[inline]
